Move all mappers into own package
