+++ server/src/main/java/org/tctalent/server/model/db/mapper/LanguageMapper.java
@@ -14,10 +14,11 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-package org.tctalent.server.model.db;
+package org.tctalent.server.model.db.mapper;
 
 import org.mapstruct.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.tctalent.server.model.db.Language;
 import org.tctalent.server.service.db.LanguageService;
 
 /**
