+++ server/src/main/java/org/tctalent/server/model/db/mapper/UserMapper.java
@@ -14,9 +14,10 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-package org.tctalent.server.model.db;
+package org.tctalent.server.model.db.mapper;
 
 import org.mapstruct.Mapper;
+import org.tctalent.server.model.db.User;
 
 /**
  * User related mappings.
